fixup! [RPC] bumpfee
Fix bumpfee segwit compatibility.

There were two problems causing bumpfee not to work with segwit transactions.
The first was that the CTxIn::scriptWitness field was not being saved after
signing.  The second was that the TransactionSignatureCreator constructor was
being called with bad arguments (with SIGHASH_ALL passed as an amount argument
instead of the hashtype argument).

Also remove mention of fallbackfee from bumprpc documentation, and throw an
error if bumpfee is called with incompatible totalFee and confTarget options.

Alex Morcos <morcos@chaincode.com> found the first segwit transaction bug and
suggested the documentation and option changes.